chore: moved constant
